testing/web-platform/tests/notifications/instance.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Notification instance basic tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/custom-data.js"></script>
<script>
const n = new Notification("Radio check",
    {
        dir: "ltr",
        lang: "aa",
        body: "This is a radio check.",
        tag: "radio_check999",
        icon: `${location.origin}/icon.png`,
        data: fakeCustomData,
    }
)
n.onshow = function() {
    n.close()
}
test(function() {
    assert_true(n instanceof Notification)
},"Notification instance exists.")
test(function() {
    assert_true("close" in n)
},"Attribute exists: close")
test(function() {
    assert_true("onclick" in n)
},"Attribute exists: onclick")
test(function() {
    assert_true("onshow" in n)
},"Attribute exists: onshow")
test(function() {
    assert_true("onerror" in n)
},"Attribute exists: onerror")
test(function() {
    assert_true("onclose" in n)
},"Attribute exists: onclose")
test(function() {
    assert_equals(n.title, "Radio check")
},"Attribute exists with expected value: title")
test(function() {
    assert_equals(n.dir, "ltr")
},"Attribute exists with expected value: dir")
test(function() {
    assert_equals(n.lang, "aa")
},"Attribute exists with expected value: lang")
test(function() {
    assert_equals(n.body, "This is a radio check.")
},"Attribute exists with expected value: body")
test(function() {
    assert_equals(n.tag, "radio_check999")
},"Attribute exists with expected value: tag")
test(function() {
    assert_equals(n.icon, `${location.origin}/icon.png`)
},"Attribute exists with expected value: icon")
test(function() {
    assert_custom_data(n.data);
},"Attribute exists with expected value: data")
</script>
